refactor(Chapter): ChaptersGetAdapter
챕터 전체 조회를 위한 단일 책임 적용 및 클린 아키텍처 적용 Adapter
