--- src/main/java/com/space/server/domain/quiz/service/implementation/QuizCreator.java
@@ -1,17 +1,0 @@
-package com.space.server.domain.quiz.service.implementation;
-
-import com.space.server.domain.quiz.domain.Quiz;
-import com.space.server.domain.quiz.domain.repository.QuizRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class QuizCreator {
-
-  private final QuizRepository quizRepository;
-
-  public void create(Quiz quiz) {
-    quizRepository.save(quiz);
-  }
-}
