--- src/main/java/com/space/server/domain/quiz/service/implementation/QuizDeleter.java
@@ -1,17 +1,0 @@
-package com.space.server.domain.quiz.service.implementation;
-
-import com.space.server.domain.quiz.domain.Quiz;
-import com.space.server.domain.quiz.domain.repository.QuizRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class QuizDeleter {
-
-  private final QuizRepository quizRepository;
-
-  public void delete(Quiz quiz) {
-    quizRepository.delete(quiz);
-  }
-}
